Switches index type to std::size_t in RS simulation

+++ tests/modules/RS/test_simulation.cpp
@@ -115,7 +115,7 @@ TEST_CASE( "Test basic RS simulation semantics", "[RS][Simulation]") {
             auto particle = std::make_unique<RS::ReactiveParticle>(&Cluster_1);
             particle->setChargeElementary(i);
             particles.push_back(std::move(particle));
-            sim.addParticle(particles[i].get(),static_cast<int>(i));
+            sim.addParticle(particles[i].get(), i);
         }
 
         sim.addParticle(p2.get(), 200);
